+++ src/mdns_beacon/beacon.py
@@ -4,11 +4,14 @@ from ipaddress import IPv4Address, IPv6Address, ip_address
 from typing import Any, List, Optional, Union
 
 from slugify import slugify
+from typing_extensions import Literal
 from zeroconf import ServiceInfo
 
 from .base import BaseBeacon
 
 logger = logging.getLogger(__name__)
+
+PROTOCOL = Literal["tcp", "udp"]
 
 
 class Beacon(BaseBeacon):
@@ -35,7 +38,7 @@ class Beacon(BaseBeacon):
         addresses: Optional[List[Union[IPv4Address, IPv6Address]]] = None,
         port: int = 80,
         type_: str = "http",
-        protocol: str = "tcp",
+        protocol: PROTOCOL = "tcp",
         ttl: int = 60,
         *args: Any,
         **kwargs: Any,