+++ tests/test_cli.py
@@ -3,6 +3,7 @@ from asyncio import AbstractEventLoop
 from contextlib import ExitStack as does_not_raise
 from ipaddress import IPv4Address, IPv6Address
 from typing import ContextManager, List, Optional, Union
+from uuid import uuid4
 
 import pytest
 from click.exceptions import BadParameter
@@ -10,7 +11,8 @@ from click.testing import CliRunner
 from pytest_mock import MockerFixture
 
 import mdns_beacon
-from mdns_beacon.cli import IpAddressParamType, main
+from mdns_beacon.cli.main import main
+from mdns_beacon.cli.types import IpAddress
 
 from helpers.contextmanager import raise_keyboard_interrupt
 
@@ -23,7 +25,7 @@ from helpers.contextmanager import raise_keyboard_interrupt
         (["--version"], f"main, version { mdns_beacon.__version__ }\n"),
     ],
 )
-def test_command_line_misc(options: List[str], expected: str) -> None:
+def test_command_line_interface(options: List[str], expected: str) -> None:
     """Test the CLI."""
     runner = CliRunner()
     result = runner.invoke(main, options)
@@ -43,7 +45,7 @@ def test_ip_address_param_type(
     address: str, raises: ContextManager, expected: Optional[Union[IPv4Address, IPv6Address]]
 ) -> None:
     """Test ip address param type."""
-    ptype = IpAddressParamType()
+    ptype = IpAddress()
     with raises:
         ip = ptype.convert(address, None, None)
         assert ip == expected
@@ -53,9 +55,12 @@ def test_ip_address_param_type(
 @pytest.mark.parametrize(
     "options,expected",
     [
-        (["example"], "Shutting down ...\n"),
+        (["example", "--protocol", "tcp"], "Shutting down ...\n"),
         (["example", "--alias", "sub1.example"], "Shutting down ...\n"),
-        (["example", "--alias", "sub1.example", "--address", "127.0.0.1"], "Shutting down ...\n"),
+        (
+            ["example", "--alias", "sub1.example", "--address", "127.0.0.1", "--type", "http"],
+            "Shutting down ...\n",
+        ),
         (
             ["example", "--alias", "sub1.example", "--address", "127.0.0.1", "--address", "::1"],
             "Shutting down ...\n",
@@ -66,8 +71,11 @@ def test_blink(
     mocker: MockerFixture, safe_loop: AbstractEventLoop, options: List[str], expected: str
 ) -> None:
     """Test beacon blink."""
-    runner = CliRunner()
+    # Ugly hack to prevent collisions during parallel tests
+    uuid = uuid4()
+    options = [opt.replace("example", f"example-{uuid}") for opt in options]  # TODO: Fix me
 
+    runner = CliRunner()
     with raise_keyboard_interrupt(timeout=6):
         result = runner.invoke(main, ["blink"] + options)