+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Changed
+- Blink cli layout.
+- Listen cli layout.
 
 ## [0.1.0] - 2021-01-18
 ### Added